milestone_2.py.
The first milestone of the Hangman task. milestone_2 defines a list of my favourite fruit, as an example, picks a random choice and prints it from that list. 
milestone_2 also includes an input function which verifies if the input is both a singular character and a character from the alphabet. It then returns if it was a valid input or not.